Added Timer Mechanic and Template levels 7-10
Added a timer that displays in the top left corner of the screen right
under the score. Updated the score given lowered it from 5000 to 500
also added the ability to make picking up diamond worth something
(currently set to 1500). Updated a bunch of the prefabs such as
collect_Diamond because they are now triggers so they don't need
rigidbody on them any more. Added levels 6-10 to the build. Gave a
default layout to levels 7 - 10. Removed some old comments out code from
a few of the scripts.