Finished designing level 9
Added a new floor type and made it a prefab. Level 9 is designed to show
off the new floor type. The floor type is Moving_Floor it moves left and
right based on a speed and a direction. Updated movement,
bullet_destory_self, and destory_self_diamond to deal with the new
platform